# [PERF] Avoid duplicated requests on admin endpoints