--- a/api/app/controllers/mno_enterprise/jpi/v1/admin/organizations_controller.rb
+++ b/api/app/controllers/mno_enterprise/jpi/v1/admin/organizations_controller.rb
@@ -8,7 +8,7 @@ module MnoEnterprise
       @organizations = @organizations.skip(params[:offset]) if params[:offset]
       @organizations = @organizations.order_by(params[:order_by]) if params[:order_by]
       @organizations = @organizations.where(params[:where]) if params[:where]
-      @organizations = @organizations.all
+      @organizations = @organizations.all.fetch
 
       response.headers['X-Total-Count'] = @organizations.metadata[:pagination][:count]
     end