--- a/api/app/controllers/mno_enterprise/jpi/v1/admin/users_controller.rb
+++ b/api/app/controllers/mno_enterprise/jpi/v1/admin/users_controller.rb
@@ -8,7 +8,7 @@ module MnoEnterprise
       @users = @users.skip(params[:offset]) if params[:offset]
       @users = @users.order_by(params[:order_by]) if params[:order_by]
       @users = @users.where(params[:where]) if params[:where]
-      @users = @users.all
+      @users = @users.all.fetch
 
       response.headers['X-Total-Count'] = @users.metadata[:pagination][:count]
     end