`@users.all` is a ProxyObject, so code like  `@users.metadata if
@users.respond_to?(:metadata)` will trigger 2 requests.
By calling `#fetch` we get a relation and only fetch it once.

Before:
```
Started GET "/mnoe/jpi/v1/admin/users?limit=10&offset=0&order_by=created_at.desc" for 127.0.0.1 at 2016-11-07 20:41:31 +1100
Processing by MnoEnterprise::Jpi::V1::Admin::UsersController#index as JSON
  Parameters: {"limit"=>"10", "offset"=>"0", "order_by"=>"created_at.desc"}
[localhost] GET /api/mnoe/v1/users/1235 (0.034 s)
[localhost] GET /api/mnoe/v1/users?limit=10&skip=0&sort%5B%5D=created_at.desc (0.025 s)
[localhost] GET /api/mnoe/v1/users?limit=10&skip=0&sort%5B%5D=created_at.desc (0.021 s)
[localhost] GET /api/mnoe/v1/users?limit=10&skip=0&sort%5B%5D=created_at.desc (0.020 s)
[localhost] GET /api/mnoe/v1/users?limit=10&skip=0&sort%5B%5D=created_at.desc (0.030 s)
[localhost] GET /api/mnoe/v1/users?limit=10&skip=0&sort%5B%5D=created_at.desc (0.031 s)
[localhost] GET /api/mnoe/v1/users?limit=10&skip=0&sort%5B%5D=created_at.desc (0.027 s)
Completed 200 OK in 1139ms (Views: 896.5ms | ActiveRecord: 0.0ms)
```

After:
```
Started GET "/mnoe/jpi/v1/admin/users?limit=10&offset=0&order_by=created_at.desc" for 127.0.0.1 at 2016-11-07 20:42:22 +1100
Processing by MnoEnterprise::Jpi::V1::Admin::UsersController#index as JSON
  Parameters: {"limit"=>"10", "offset"=>"0", "order_by"=>"created_at.desc"}
[localhost] GET /api/mnoe/v1/users/1235 (0.034 s)
[localhost] GET /api/mnoe/v1/users?limit=10&skip=0&sort%5B%5D=created_at.desc (0.023 s)
Completed 200 OK in 263ms (Views: 33.0ms | ActiveRecord: 0.0ms)
```

@alexnoox FYI